+++ include/math/Vector3.hpp
@@ -7,8 +7,8 @@ struct Vector3
 {
    // CONSTRUCTORS
    Vector3();
-   Vector3(double in);
-   Vector3(double _x, double _y, double _z);
+   Vector3(float in);
+   Vector3(float _x, float _y, float _z);
    Vector3(const Vector3& other);
 
    ~Vector3();
@@ -19,29 +19,29 @@ struct Vector3
    Vector3 &operator=(const Vector3& other);
 
    Vector3 operator+(const Vector3& other);
-   Vector3 operator+(double a);
+   Vector3 operator+(float a);
 
    Vector3 operator-(const Vector3& other);
-   Vector3 operator-(double a);
+   Vector3 operator-(float a);
 
-   double  operator*(const Vector3& other); // Dot
+   float  operator*(const Vector3& other); // Dot
    Vector3 operator*(const Matrix3& other);
-   Vector3 operator*(double a);             // Element Wise Scalar
+   Vector3 operator*(float a);             // Element Wise Scalar
 
    Vector3 operator/(const Vector3& other);
-   Vector3 operator/(double a);
+   Vector3 operator/(float a);
 
    Vector3 operator^(const Vector3& other); // Cross
 
    // DATA
-   double x;
-   double y;
-   double z;
+   float x;
+   float y;
+   float z;
 };
 
 // UTILITY FUNCTIONS
-double Mag(Vector3 a);
-double Dot(Vector3 a, Vector3 b);
+float Mag(Vector3 a);
+float Dot(Vector3 a, Vector3 b);
 Vector3 ElementMult(Vector3 a, Vector3 b);
 Vector3 Cross(Vector3 a, Vector3 b);
 Vector3 Proj(Vector3 a, Vector3 b);
